Add Jest and update to TypeScript 3.0.1
