+++ tsconfig.json.old
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "module": "commonjs",
+    "target": "es5",
+    "noImplicitAny": true,
+    "sourceMap": true,
+    "rootDir": "./src",
+    "outDir": "./",
+    "declaration": true,
+    "listFiles": true,
+    "newLine": "LF",
+    "pretty": true
+  },
+  "include": [
+    "src/**/*"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
+}
